Shipment stage

+++ Models/TransportRoute.cs
@@ -10,7 +10,7 @@ namespace DHLManagementSystem.Models
 
         public string? Destination { get; set; }
 
-        public double DistanceKm { get; set; }
+        public decimal DistanceKm { get; set; }
 
         public decimal BasePrice { get; set; }
     }
